// frontend/src/client/models/ProjectRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProjectRead = {
    id: string;
    client_id: string;
    name: string;
    url: string;
    description?: (string | null);
    sitemap_url?: (string | null);
    created_at: string;
};